UPD: Eeg_processor's Dockerfile to use ROS2 Humble
Due to the Docker base image using Galactic not being able to successfully
install Psychopy, for a reason unknown to me. (It used to work earlier.)

However, we are planning to upgrade to Humble in any case, so here it is
done for eeg_processor first.

Note that until the rest of the containers are updated, communication
between the containers most likely doesn't work.

Downgrading NumPy version doesn't seem to be required anymore, hence
ditch that line in Dockerfile.