DEL: Dockerfile.prod versions of Dockerfiles, and related scripts
The idea giving rise to Dockerfile.prod files is that we
should run Docker containers on the other installation sites
(Tubingen, Chieti) without exposing the source code to the
sites. Dockerfile.prod files are variants of Dockerfiles which
differ in that they delete the source code from the containers
after the containers were built. However:

- We cannot have a fully Dockerized production environment with
the current technology stack because the FPGA cannot be operated
over Docker.

- With C++ nodes, it is feasible to not expose the source code
on the sites, as the C++ nodes are built into binary files.
However, with Python nodes, we cannot easily hide the source
code from being accessed inside the container.

- If we cannot implement the feature in any case, we should not
maintain parts of the code whose sole purpose is to implement
that feature.

For those reasons, delete the production versions of Dockerfiles
and the related scripts for now. They can always be re-introduced
later if desired.