MOD: Move internally to 0-based indexing for channels
Everywhere else than in the FPGA; LabVIEW has one-based indexing,
hence the different convention there. Transform zero-based indexing
to one-based indexing in mTMS device bridge for now. However, the
correct place for that would be on the FPGA.

Only use 1-based indexing for channels in the user-facing parts,
namely, the UI.