MOD: API: HACK: Discharge to minimum of 3 volts
This is to prevent long discharge times, which can occur
when discharging to 0-2, preventing, e.g., efficient use
of coil characterizer script, as it needs to repeatedly
charge and discharge to given target voltages in fast
succession.

Work around by always discharging to a minimum of 3 volts
in both Python and MATLAB APIs. Mark the hacky parts of
code with XXX comment.