FIX: Major safety issue related to Decider
Previously, experiment in mTMS panel could be run when Decider
was on, causing the Decider (which uses PWM pulses) to charge
all capacitors to the maximum voltage, while the mTMS panel
experiments use non-PWM pulses. This caused potentially very
high-intensity pulses to be delivered when an experiment was
performed via the mTMS panel if the Decider was on simultaneously.

Prevent this by disabling "Start" button in Experiment panel
if Decider is enabled.